pyproject.toml
[project]
name = "graph-reader"
version = "0.1.0"
description = ""
authors = [{ name = "Ali Sedaghatbaf", email = "ali.sedaghatbaf@klarna.com" }]
requires-python = "~=3.11"
readme = "README.md"
dependencies = [
    "klarna-wiki-api>=0.12,<0.13",
    "neo4j>=5.26.0,<6",
    "langchain-core>=0.3.17,<0.4",
    "langchain-openai>=0.2.8,<0.3",
    "langchain-text-splitters>=0.3.2,<0.4",
    "tiktoken>=0.8.0,<0.9",
    "wikipedia>=1.4.0,<2",
    "pandas>=2.2.3,<3",
    "pydantic>=2.9.2,<3",
    "langgraph>=0.2.46,<0.3",
    "beautifulsoup4>=4.12.3,<5",
    "flake8>=7.1.1,<8",
    "black>=24.10.0,<25",
    "isort>=5.13.2,<6",
    "numpy==1.26.2",
    "langchain-community>=0.3.8,<0.4",
    "sparqlwrapper>=2.0.0,<3",
    "python-dotenv>=1.0.1,<2",
    "mypy>=1.13.0,<2",
    "watchdog>=6.0.0,<7",
    "streamlit>=1.41.0,<2",
    "langsmith>=0.2.3,<0.3",
    "torch>=2.5.1,<3",
    "docling>=2.12.0,<3",
    "docling-core>=2.10.0,<3",
    "langchain-huggingface>=0.1.2,<0.2",
    "langchain-milvus>=0.1.7,<0.2",
    "langchain-experimental>=0.3.3,<0.4",
    "langgraph-cli[inmem]>=0.1.65,<0.2",
    "sentence-transformers>=3.3.1,<4",
    "langgraph-checkpoint-sqlite>=2.0.1,<3",
    "keybert>=0.8.5,<0.9",
    "scikit-learn>=1.6.1,<2",
    "rank-bm25>=0.2.2,<0.3",
    "selenium>=4.29.0",
    "pymupdf>=1.25.3",
    "layoutparser>=0.3.4",
]

[dependency-groups]
dev = ["autoflake>=2.3.1,<3"]

[tool.uv]

[[tool.uv.index]]
name = "klarna-proprietary"
url = "https://artifactory.klarna.net/artifactory/api/pypi/v-pypi-production/simple"
explicit = true

[[tool.uv.index]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
klarna-wiki-api = { index = "klarna-proprietary" }




